fix: More meaningful error when creating a datadoc in env without query engine
