+++ querybook/webapp/components/CreateDataDocButton/CreateDataDocButton.tsx
@@ -1,14 +1,7 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-
-import history from 'lib/router-history';
 import { TooltipDirection } from 'const/tooltip';
-import * as dataDocActions from 'redux/dataDoc/action';
-import { currentEnvironmentSelector } from 'redux/environment/selector';
-import { Dispatch, IStoreState } from 'redux/store/types';
+import { useCreateDataDoc } from 'hooks/dataDoc/useCreateDataDoc';
 import { IconButton } from 'ui/Button/IconButton';
-import { getQueryEngineId } from 'lib/utils';
-import { queryEngineSelector } from 'redux/queryEngine/selector';
 
 export interface ICreateDataDocButtonProps {
     // from own Props
@@ -20,26 +13,7 @@ export const CreateDataDocButton: React.FunctionComponent<ICreateDataDocButtonPr
     tooltipPos = 'left',
     tooltip = 'New DataDoc',
 }) => {
-    const environment = useSelector(currentEnvironmentSelector);
-    const queryEngines = useSelector(queryEngineSelector);
-    const defaultEngineId = useSelector((state: IStoreState) =>
-        getQueryEngineId(
-            state.user.computedSettings['default_query_engine'],
-            queryEngines.map(({ id }) => id)
-        )
-    );
-    const dispatch: Dispatch = useDispatch();
-
-    const handleCreateDataDoc = React.useCallback(() => {
-        const cell = {
-            type: 'query',
-            context: '',
-            meta: { engine: defaultEngineId },
-        };
-        dispatch(dataDocActions.createDataDoc([cell])).then((dataDoc) =>
-            history.push(`/${environment.name}/datadoc/${dataDoc.id}/`)
-        );
-    }, [environment]);
+    const handleCreateDataDoc = useCreateDataDoc();
 
     return (
         <IconButton
